vpp-example-plugin: add a skeleton of an example plugin
Add an example plugin that just registers itself at the moment. This
generates a C dylib (.so) for use with VPP.

Add a Python test framework (derived from that in VPP) to allow
running integration tests.

Add workspace-wide profile options to generate smaller, faster release
code ready for when the plugin is used as a forwarding node.