Add vlib_init_function macro
Add support for the vlib_init_function proc-macro to be applied to a
function to allow it to be invoked by VPP at the appropriate time at
startup, making use of the ctor crate to avoid needing to reimplement
constructor/destructor attribute support, which is non-trivial,
especially across different targets.

Add abstractions for vlib_main_t, including one for with the barrier
held which will be used later to catch incorrect uses of APIs through
the Rust type system.