Feature registration
Add support for feature registration to allow using a node as part of
a feature arc.

This is done through adding a new vnet_feature_init proc macro that
assists with constructing the C structure for the feature and
registering the feature at init time.

This is done separate from the node registration following what VPP
does and allowing for the same feature to be registered in different
feature arcs (assuming that the node function preconditions apply to
those feature arcs). Unlike with vpp though, a strong dependency is
established with the node, as opposed to a weak dependency with just
using a string. This is for two reasons:
1. To avoid accidental mismatches between the registered node name and
   node name in the feature registration.
2. To allow zero-cost validations enforced by the type system,
   currently limited to the node FeatureData and ensuring this is
   consistent between the feature enablement and the feature get next
   call during node processing.

An newtype wrapper is added for SwIfIndex for type-safety and for
convenience, allowing FromStr/Display to be implemented.