API generator
Add a parser for VPP .api files and a generator for both Rust code and
JSON. The JSON is used by the Rust code to register the JSON
representation with VPP for retrieval via an API.

Although the parser is considered complete, the generator is less so,
although the unimplemented features should error rather than
generating code that doesn't compile.

Although it would be nice to implement Deserialize for the requests
from a byte stream, the VPP APIs don't make this easy since there is
no one callback that provides both the bytes and the message
length. So instead, the generated types are expected to have the same
layout as the on wire representation (except for endianness).