vec: apply no allocation on creation optimisation
Both VPP vectors and Rust standard library vectors have an
optimisation whereby they avoid allocating on creation of an empty
vector, delaying allocation to when the first element is added.

Therefore, improve compatibility with VPP vectors (whereby we might be
taking ownership of a C pointer using Vec::from_raw) by supporting the
Vec being a null pointer, and applying the same optimisation such that
Vec::new just initialises as a null pointer.

Note that Rust standard library vectors use a dangling pointer
instead, which can be done because the length isn't held in the
pointer. However, for VPP vectors, the length is held in a header
immediately before the pointer to the first element, so a dangling
pointer wouldn't work for VPP vectors (even ignoring the compatibility
requirement).

One important consequence is that we have to remove the Deref/Borrow
trait implementations for Vec to VecRef since it isn't allowed to
create a reference (to a VecRef) from a null pointer. An alternative,
would be to allocate upon the first Deref/Borrow, but that would be
surprising behaviour and we don't have a mutable self in some cases so
we'd be off into the weeds. The removal of the Deref trait
implementation means that we now need to duplicate method declarations
from VecRef in Vec.